+++ bitvmx_protocol_library/transaction_generation/publication_services/prover/publish_hash_transaction_service.py
