+++ bitvmx_protocol_library/transaction_generation/publication_services/verifier/publish_choice_search_transaction_service.py
