+++ bitvmx_protocol_library/transaction_generation/publication_services/verifier/trigger_execution_challenge_transaction_service.py
