+++ bitvmx_protocol_library/transaction_generation/publication_services/verifier/trigger_wrong_hash_challenge_transaction_service.py
