+++ bitvmx_protocol_library/transaction_generation/verifier_challenge_detection/verifier_execution_challenge_detection_service.py
@@ -1,6 +1,6 @@
 from bitvmx_execution.services.execution_trace_query_service import ExecutionTraceQueryService
 from bitvmx_protocol_library.transaction_generation.enums import TransactionVerifierStepType
-from bitvmx_protocol_library.transaction_generation.publication_services.trigger_execution_challenge_transaction_service import (
+from bitvmx_protocol_library.transaction_generation.publication_services.verifier.trigger_execution_challenge_transaction_service import (
     TriggerExecutionChallengeTransactionService,
 )
